+++ src/Microsoft.DotNet.Wpf/test/DRT/DrtXaml/BamlTestClasses40/BamlTestClasses40.csproj
@@ -62,7 +62,8 @@
     <Page Include="XmlSpace3.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
-
+    <Compile Include="BCMarkupExtension.cs" />
+    <Compile Include="TempDPSetOrder1ListBox.cs" />
     <None Remove="@(Page)" />
   </ItemGroup>
