+++ src/Microsoft.DotNet.Wpf/test/DRT/DrtXaml/BamlTestClasses40/TempDPSetOrder1.xaml.cs
@@ -16,47 +16,4 @@ namespace BamlTestClasses40
             InitializeComponent();
         }
     }
-
-    public enum Property
-    {
-        IsSynchronizedWithCurrentItem,
-        ItemsSource,
-        SelectedIndex
-    }
-
-    public class TempDPSetOrder1ListBox : ListBox
-    {
-        public static List<Property> Order { get; private set; }
-
-        static TempDPSetOrder1ListBox()
-        {
-            Order = new List<Property>();
-        }
-
-        protected override void OnPropertyChanged(DependencyPropertyChangedEventArgs e)
-        {
-            base.OnPropertyChanged(e);
-            if (e.Property.OwnerType.Equals(typeof(Selector)))
-            {
-                switch (e.Property.Name)
-                {
-                    case "IsSynchronizedWithCurrentItem":
-                        Order.Add(Property.IsSynchronizedWithCurrentItem);
-                        break;
-                    case "SelectedIndex":
-                        Order.Add(Property.SelectedIndex);
-                        break;
-                }
-            }
-            else if (e.Property.OwnerType.Equals(typeof(ItemsControl)))
-            {
-                switch (e.Property.Name)
-                {
-                    case "ItemsSource":
-                        Order.Add(Property.ItemsSource);
-                        break;
-                }
-            }
-        }
-    }
 }
